requirements.txt
langchain==0.0.234
torch==2.2.1
accelerate==0.27.2
pypdf2
python-dotenv
transformers==4.38.1
sentence_transformers==2.4.0
faiss-cpu==1.7.4
streamlit==1.31.0
streamlit-chat==0.1.1
tiktoken
huggingface-hub